change from float to int

+++ ch01/exercises/main.py
@@ -6,8 +6,8 @@ print(myvalues)
 rate = input("Current Exchange Rate of USD to Euros: ")
 amount = input("Amount to Exchange: ")
 
-rate = float(rate)
-amount = float(amount)
+rate = int(rate)
+amount = int(amount)
 total = amount*rate
 servicefee = 3.00
 result = total - servicefee
